+++ src/DIBSearchPopup.hpp
@@ -1,9 +1,30 @@
 #include "DemonsInBetween.hpp"
 
+class TableNode : public CCNode {
+protected:
+    CCArray* m_menus;
+    AxisLayout* m_columnLayout;
+    AxisLayout* m_rowLayout;
+    int m_columns;
+    int m_rows;
+    float m_rowHeight;
+
+    bool init(int columns, int rows);
+public:
+    static TableNode* create(int columns, int rows);
+
+    void setColumnLayout(AxisLayout*);
+    void setRowLayout(AxisLayout*);
+    void setRowHeight(float);
+    void updateAllLayouts();
+    void addButton(CCMenuItemSpriteExtra*);
+
+    ~TableNode() override;
+};
+
 class DIBSearchPopup : public Popup<> {
 protected:
     bool setup() override;
-    void createDifficultyButton(CCMenu*, int);
 public:
     static DIBSearchPopup* create();
 };